+++ blender-qmm/mats/base/Zinc.py
@@ -24,7 +24,7 @@ class QMMZinc(bpy.types.Operator):
             #CreateShader
             m_zinc = bpy.data.materials.new(name = "QMM Zinc")
             m_zinc.use_nodes = True
-            m_zinc.diffuse_color = (0.672442, 0.715693, 0.723055, 1)
+            m_zinc.diffuse_color = (0.737911, 0.723055, 0.701102, 1)
             m_zinc.metallic = 1
             m_zinc.roughness = 0.3
 
@@ -37,8 +37,7 @@ class QMMZinc(bpy.types.Operator):
             #princibledbsdf
             BSDF = nodes.get('Principled BSDF')
             BSDF.location = (-300,0)
-            # BSDF.inputs[0].default_value = (0.23074, 0.242281, 0.250158, 1)
-            BSDF.inputs[0].default_value = (0.912, 0.914, 0.920, 1)
+            BSDF.inputs[0].default_value = (0.737911, 0.723055, 0.701102, 1)
             BSDF.inputs[6].default_value = 1
             BSDF.inputs[9].default_value = 0.3
             BSDF.inputs[16].default_value = 2.368
@@ -49,7 +48,7 @@ class QMMZinc(bpy.types.Operator):
             ec_group.node_tree = bpy.data.node_groups['Energy Conservation']
             ec_group.location = (-500, -200)
             ec_group.inputs[0].default_value = 2.368
-            ec_group.inputs[1].default_value = (0.672442, 0.715693, 0.723055, 1)
+            ec_group.inputs[1].default_value = (0.737911, 0.723055, 0.701102, 1)
             ec_group.inputs[2].default_value = (0.01, 0.01, 0.01, 1)
 
             links = m_zinc.node_tree.links.new